move promise-related errors next to the promise itself
