+++ test/unittest/presentation/test_client.cpp
@@ -18,7 +18,6 @@
 #include <cetl/pf17/cetlpf.hpp>
 #include <libcyphal/errors.hpp>
 #include <libcyphal/presentation/client.hpp>
-#include <libcyphal/presentation/errors.hpp>
 #include <libcyphal/presentation/presentation.hpp>
 #include <libcyphal/presentation/response_promise.hpp>
 #include <libcyphal/transport/errors.hpp>
